refactor: 계정, 고객지원 컴포넌트 client 단으로 분리 및 파일 경로 수정 (#21)
