+++ apps/web/src/app/mypage/_components/AccountInfoContainer/AccountInfoContainer.tsx
@@ -1,0 +1,10 @@
+import AccountInfoClient from './_clientBoundary/AccountInfoClient';
+import styles from './AccountInfoContainer.module.css';
+
+export default function AccountInfoContainer() {
+  return (
+    <section className={styles.wrapper}>
+      <AccountInfoClient />
+    </section>
+  );
+}
