+++ apps/web/src/app/mypage/_components/CustomerSupportContainer/CustomerSupportContainer.module.css
@@ -1,0 +1,6 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  padding: 0 16px;
+}
