+++ apps/web/src/app/mypage/_components/CustomerSupportContainer/CustomerSupportContainer.tsx
@@ -1,0 +1,10 @@
+import CustomerSupportClient from './_clientBoundary/CustomerSupportClient';
+import styles from './CustomerSupportContainer.module.css';
+
+export default function CustomerSupportContainer() {
+  return (
+    <section className={styles.wrapper}>
+      <CustomerSupportClient />
+    </section>
+  );
+}
