+++ apps/web/src/app/mypage/page.module.css
@@ -25,6 +25,14 @@
   margin-top: 28px;
 }
 
+.accountInfo {
+  margin-top: 28px;
+}
+
+.customerSupport {
+  margin-top: 28px;
+}
+
 .divider {
   margin-top: 32px;
 }
